requirements.txt
flask
flask-cors
torch
torchvision
numpy
scikit-learn
pillow
